Added BPF_PROG_VERIFY and BPF_PROG_EXTRACT commands to the bpf() syscall.
BPF_PROG_VERFIY will be used to verify a bpf program. It will return the
number of bytes that we need to sign and copy out to user space. This
allows a user loading program to allocate enough memory to store the
whole object. I'm not sure if this is the best way to do this but it is
a start. I think there is some way to mmap() but I'm not sure.

BPF_PROG_EXTRACT will extract the data out of the kernel into a user
space pointer.