Adding support for using extract_len field
If you want to use the extract_len field to know how much data
to copy out you can now use the bpf_object__set_extract_len() function
to set that. bpf_object__load() will now copy that in to the attr and it
should get populated with the right length.