BPF_PROG_LOAD_VERIFIED
For bpf native code allocation there is a special way to allocate it.
I think this makes the code RO+X (read only and execute).
It also places it in the right segment of memory to call the kernel
functions etc etc.

The allocator is bpf_prog_pack.
We use the functions bpf_jit_binary_pack_alloc and
bpf_jit_binary_pack_finalize.

The first creates the memory and gives us a rw pointer to use.
We write our code to that and then call the second function to copy it
into the ro pointer.

There is some architecture specific shenanigans going on here to figure
out.