BPF_PROG_EXTRACT: Extract helper name
In an effort to standardize bpf program loading, we are switching to
extracting the name of the helper funciton. This will allow for kfuncs
to be reloacted, as well as using kallsysms to lookup the symbol name to
get the address, instead of relying on the bpf_helper id. This should be
more general.