Add per process BPF tracing support
Added a field in the task struct to indicate if bpf tracing should be
enabled or disabled. Currently disabled by default.

Added BPF_PROCESS_{ENABLE,DISABLE} options to the bpf syscall to change
the status of the bpf_enable field for the current process.

Added a check in trace_call_bpf to see if bpf tracing is enabled for the
process that triggered the event.